requirements.txt
numpy
tf-keras
faiss-cpu
deepface
pillow